feat: update sanity structure
